+++ src/components/templates/TodoPage/index.tsx
@@ -1,35 +1,18 @@
-import React, { useState } from 'react';
+import React from 'react';
+import useTodo from '../../../hooks/useTodo'; // カスタムフックをインポート
 import TodoForm from '../../organisms/TodoForm';
 import TodoList from '../../organisms/TodoList';
 import Input from '../../atoms/Input';
 import styles from './style.module.css';
 
 const TodoPage: React.FC = () => {
-  const [todos, setTodos] = useState<{ id: number; title: string }[]>([]);
-  const [search, setSearch] = useState('');
-
-  // Todoを追加する関数
-  const handleAddTodo = (title: string) => {
-    setTodos(prevTodos => [
-      ...prevTodos,
-      { id: prevTodos.length + 1, title },
-    ]);
-  };
-
-  // Todoを削除する関数
-  const handleDeleteTodo = (id: number) => {
-    setTodos(prevTodos => prevTodos.filter(todo => todo.id !== id));
-  };
-
-  // 検索の入力を処理する関数
-  const handleSearchChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearch(e.target.value);
-  };
-
-  // 検索結果に基づいてTodoをフィルタリング
-  const filteredTodos = todos.filter(todo => 
-    todo.title.toLowerCase().startsWith(search.toLowerCase())
-  );
+  const {
+    search,
+    filteredTodos,
+    addTodo,
+    deleteTodo,
+    handleSearchChange,
+  } = useTodo();
 
   return (
     <div className={styles.todoPage}>
@@ -39,8 +22,8 @@ const TodoPage: React.FC = () => {
         onChange={handleSearchChange}
         placeholder="検索"
       />
-      <TodoForm onAddTodo={handleAddTodo} />
-      <TodoList todos={filteredTodos} onDelete={handleDeleteTodo} />
+      <TodoForm onAddTodo={addTodo} />
+      <TodoList todos={filteredTodos} onDelete={deleteTodo} />
     </div>
   );
 };
